Move conductor into its own directory
