+++ arangod/Pregel/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_library(arango_pregel STATIC
   AggregatorHandler.cpp
   AlgoRegistry.cpp
-  Conductor.cpp
   ExecutionNumber.cpp
   GraphStore.cpp
   IncomingCache.cpp
@@ -26,4 +25,5 @@ target_include_directories(arango_pregel PRIVATE
   "${PROJECT_SOURCE_DIR}/${ENTERPRISE_INCLUDE_DIR}")
 
 add_subdirectory(Algos)
+add_subdirectory(Conductor)
 add_subdirectory(REST)
