+++ arangod/Pregel/Conductor/CMakeLists.txt
@@ -1,0 +1,2 @@
+target_sources(arango_pregel PRIVATE
+  Conductor.cpp)
